Rename all `event*` modules into `message*`
No other type or variable name changes.
